Keep upstream price import out of page
